SERVER: split up previous_results and result_storage
SERVER: split up previous_results into one result related and result overview related functionality, split result_storage similarly
CLIENT: updated the requests to work with the new server route
