+++ client/src/components/Result.vue
@@ -64,7 +64,7 @@ async function getHeaderOptions(index) {
       name: props.result.result[index].dataset.dataset,
     }),
   }
-  const response = await fetch(API_URL + '/all-results/headers', requestOptions)
+  const response = await fetch(API_URL + '/result/headers', requestOptions)
   const data = await response.json()
   const headerOptions = data
   console.log(data)
@@ -86,7 +86,7 @@ async function setRecs(currentTable) {
   };
   console.log('sending to server:', requestOptions.body);
   const response = await fetch(
-    API_URL + '/all-results/set-recs',
+    API_URL + '/result/set-recs',
     requestOptions
   );
   // console.log('resultfetch', response)
@@ -108,7 +108,7 @@ async function setRecs(currentTable) {
 //     body: JSON.stringify({ id: props.result.id }),
 //   }
 //   const response = await fetch(
-//     API_URL + '/all-results/result-by-id',
+//     API_URL + '/result/result-by-id',
 //     requestOptions
 //   ).then(() => {
 //     console.log('succesful POST request to API to retrieve evaluation data')
@@ -120,7 +120,7 @@ async function setRecs(currentTable) {
 
 //GET request: Ask server for currently loaded evaluations
 async function getEvaluations() {
-  const response = await fetch(API_URL + '/all-results/result-by-id')
+  const response = await fetch(API_URL + '/result/result-by-id')
   console.log('succesfully retrieved evaluation data.')
   const resultsData = await response.json()
   console.log('results data', resultsData)
@@ -148,7 +148,7 @@ async function getUserRecs(currentTable) {
     }),
   };
 
-  const response = await fetch(API_URL + '/all-results/result', requestOptions);
+  const response = await fetch(API_URL + '/result/', requestOptions);
   data.value.results[currentTable] = await response.json();
   selectedHeaders.value[currentTable] = Object.keys(
     data.value.results[currentTable][0]
@@ -163,7 +163,7 @@ async function exportTable(currentTable) {
     body: JSON.stringify({ results: props.result.result[currentTable].results }),
   }
   const response = await fetch(
-    API_URL + '/all-results/export',
+    API_URL + '/result/export',
     requestOptions
   )
   const confirmation = await response.json()
@@ -178,7 +178,7 @@ async function validate() {
     body: JSON.stringify({ filepath: file, amount: validationAmount.value }),
   }
   const response = await fetch(
-    API_URL + '/all-results/validate',
+    API_URL + '/result/validate',
     requestOptions
   ).then(() => {
     console.log('Experiment validated succesfully')
