+++ client/src/helpers/resultRequests.js
@@ -6,7 +6,7 @@ import { API_URL } from '../api'
 import { addResult, store } from '../store'
 import { formatResult } from './resultFormatter'
 
-const resultsRoute = '/all-results/result-by-id'
+const resultsRoute = '/result/result-by-id'
 const url = API_URL + resultsRoute
 
 // Request full result from result ID (timestamp)
