+++ server/project/__init__.py
@@ -8,7 +8,7 @@ from flask_cors import CORS
 
 from .experiment import compute_bp
 from .music_detail import detail_bp
-#from .result import result_bp
+from .result import result_bp
 from .previous_results import results_bp
 from .result_storage import create_results_overview
 from .mail import make_mail, mail_bp
@@ -55,7 +55,7 @@ def register_blueprints(app):
     :param app: the main app
     """
     app.register_blueprint(compute_bp)
-    #app.register_blueprint(result_bp)
+    app.register_blueprint(result_bp)
     app.register_blueprint(results_bp)
     app.register_blueprint(detail_bp)
     app.register_blueprint(mail_bp)
