--- server/project/headers.json
@@ -1,39 +1,0 @@
-{
-  "LFM-1B": {
-    "headers": ["rating"],
-    "userHeaders": [
-      "user_country",
-      "user_age",
-      "user_gender",
-      "user_plays",
-      "user_registered"
-    ],
-    "itemHeaders": ["artist_name"]
-  },
-  "LFM-360K": {
-    "headers": ["rating"],
-    "userHeaders": [
-      "user_sha",
-      "user_gender",
-      "user_age",
-      "user_country",
-      "user_signup"
-    ],
-    "itemHeaders": ["artist_name"]
-  },
-  "ML-25M": {
-    "headers": ["rating"],
-    "itemHeaders": ["movie_title", "movie_genres"]
-  },
-  "ML-100K": {
-    "headers": ["rating"],
-    "userHeaders": ["user_age", "user_gender", "user_occupation", "user_zip code"],
-    "itemHeaders": [
-      "movie_title",
-      "movie_release date",
-      "movie_video release date",
-      "movie_imdb url",
-      "movie_genres"
-    ]
-  }
-}
