--- server/project/test_results/magic_mock.tsv
@@ -1,8 +1,0 @@
-timestamp   number  reverse
-0   0   foo
-1   1   bar
-2   2   foo
-3   3   bar
-4   4   foo
-5   5   bar
-6   6   foo
